i made the TV fly
has a few issues.

for some reason it does really quick 360s in the middle of the flying animation.

also, it has a twitch when it enters active state, both after the initial animation, and when it gets out of the bored animation

currently set to begin at 8seconds for testing